+++ gamemodes/terrortown/gamemode/client/cl_voice.lua
@@ -2,36 +2,28 @@
 -- Voicechat popup
 -- @module VOICE
 
-DEFINE_BASECLASS("gamemode_base")
-
 local GetTranslation = LANG.GetTranslation
 local string = string
 local math = math
 local net = net
 local player = player
-local pairs = pairs
-local timer = timer
 local IsValid = IsValid
 local hook = hook
-local surface = surface
+
+VOCIE_MODE_GLOBAL = 0
+VOICE_MODE_TEAM = 1
+VOICE_MODE_SPEC = 2
 
 -- voicechat stuff
 VOICE = {}
 
-local VP_GREEN = Color(0, 200, 0)
-local VP_RED = Color(200, 0, 0)
 local MutedState
-
--- voice popups, copied from base gamemode and modified
-
-g_VoicePanelList = nil
 
 VOICE.cv = {
     ---
     -- @realm client
     -- stylua: ignore
     duck_spectator = CreateConVar("ttt2_voice_duck_spectator", "0", {FCVAR_ARCHIVE}),
-
     ---
     -- @realm client
     -- stylua: ignore
@@ -59,8 +51,13 @@ end
 CreateVoiceTable()
 
 local function VoiceTryEnable()
-    if not VOICE.IsSpeaking() and VOICE.CanSpeak() and VOICE.CanEnable() then
+    local client = LocalPlayer()
+
+    if not client:IsSpeakingInVoice() and VOICE.CanSpeak() and VOICE.CanEnable() then
+        client:SetSpeakingInVoice(true)
+
         VOICE.isTeam = false
+
         permissions.EnableVoiceChat(true)
 
         return true
@@ -70,7 +67,17 @@ local function VoiceTryEnable()
 end
 
 local function VoiceTryDisable()
-    if VOICE.IsSpeaking() and not VOICE.isTeam then
+    local client = LocalPlayer()
+
+    -- when connecting to a server this function might be executed before the local
+    -- player is ready. To prevent this, we have to check if it is valid
+    if not IsValid(client) then
+        return
+    end
+
+    if not VOICE.isTeam then
+        client:SetSpeakingInVoice(false)
+
         permissions.EnableVoiceChat(false)
 
         return true
@@ -95,6 +102,91 @@ VOICE.ActivationModes = {
 }
 
 ---
+-- Generates a fake voice spectrum based on the player voice volume that looks like
+-- the restult of an FFT.
+-- @param Player ply The player that should generate a fake voice spectrum
+-- @param[default=16] number stepCount Defines the result's resolution
+-- @return table Returns the table with the fake spectrum
+-- @realm client
+function VOICE.GetFakeVoiceSpectrum(ply, stepCount)
+    stepCount = stepCount or 16
+
+    ply.lastSteps = ply.lastSteps or {}
+
+    -- at first the volume is boosted and limited so that the voice level range
+    -- makes a nice sweep over the whole spectrum
+    local volume = math.min(4, ply:VoiceVolume() * 6 + math.Rand(-0.2, 0.2))
+
+    local biggestValue = 0
+
+    for i = 1, stepCount do
+        local progress = (i - 1) / stepCount
+
+        -- the base for the fake spectrum is a simple sweep over the whole range
+        local value = 2 ^ (-(volume * 4 - 2.2 - progress * 4) ^ 2)
+
+        -- since most people talk a bit quiet, we add a signal to the end of the spectrum
+        value = value + 2 ^ (-(4.5 - 4 * progress) ^ 2) * ply:VoiceVolume() ^ 0.5
+
+        -- also the whole spectrum is linearly shifted, when the voice is louder
+        value = value + 2.5 * ply:VoiceVolume()
+
+        -- to decrease jumpyness a slow lerp is put on top which smoothes everything
+        value = Lerp(7.5 * FrameTime(), ply.lastSteps[i] or 0, value)
+
+        -- then a random value is put on top that makes it look less static and slow
+        value = value + math.Rand(-0.2 * value ^ 2, 0.2 * value ^ 2)
+        value = math.max(0, value)
+
+        -- in the last step a fast lerp is applied that only takes the edge of the random noise
+        ply.lastSteps[i] = Lerp(150 * FrameTime(), ply.lastSteps[i] or 0, value)
+
+        biggestValue = math.max(biggestValue, value)
+    end
+
+    -- make sure that the spectrum never surpasses 1; if the values are too big, the spectrum
+    -- is normalized
+    if biggestValue > 1.0 then
+        for i = 1, stepCount do
+            ply.lastSteps[i] = ply.lastSteps[i] / biggestValue
+        end
+    end
+
+    return ply.lastSteps
+end
+
+---
+-- Sets the mode of the voice panel.
+-- @param Player ply The player whose voice mode should be set
+-- @param number mode The voice mode
+-- @realm client
+function VOICE.SetVoiceMode(ply, mode)
+    ply.voiceMode = mode
+end
+
+---
+-- Returns the mode of the voice panel.
+-- @param Player ply The player whose vocie mode should be read
+-- @return number The mode of the voice panel
+-- @realm client
+function VOICE.GetVoiceMode(ply)
+    return ply.voiceMode or VOICE_MODE_GLOBAL
+end
+
+---
+-- Returns the color of the voice panel.
+-- @param Player ply The player whose voice color should be read
+-- @return Color The color of the voice panel
+-- @realm client
+function VOICE.GetVoiceColor(ply)
+    if VOICE.GetVoiceMode(ply) == VOICE_MODE_GLOBAL then
+        return INNOCENT.color
+    else
+        return TEAMS[ply:GetTeam()].color
+    end
+end
+
+---
 -- Creates a closure that dynamically calls a function from VOICE.ActivationModes depending on the current mode.
 -- @param string func The name of the function to call on the current voice activation mode
 -- @return function A closure that calls the function on the current voice activation mode, if it exists
@@ -111,7 +203,11 @@ end
 hook.Add("TTT2FinishedLoading", "TTT2ActivateVoiceChat", VOICE.ActivationModeFunc("OnJoin"))
 
 local function VoiceTeamTryEnable()
-    if not VOICE.IsSpeaking() and VOICE.CanSpeak() and VOICE.CanTeamEnable() then
+    local client = LocalPlayer()
+
+    if not client:IsSpeakingInVoice() and VOICE.CanSpeak() and VOICE.CanTeamEnable() then
+        client:SetSpeakingInVoice(true)
+
         VOICE.isTeam = true
 
         permissions.EnableVoiceChat(true)
@@ -123,7 +219,11 @@ local function VoiceTeamTryEnable()
 end
 
 local function VoiceTeamTryDisable()
-    if VOICE.IsSpeaking() and VOICE.isTeam then
+    local client = LocalPlayer()
+
+    if VOICE.isTeam then
+        client:SetSpeakingInVoice(false)
+
         permissions.EnableVoiceChat(false)
 
         return true
@@ -201,32 +301,6 @@ bind.Register(
     KEY_T
 )
 
--- 255 at 100
--- 5 at 5000
-local function VoiceNotifyThink(pnl)
-    local client = LocalPlayer()
-
-    if
-        not IsValid(pnl)
-        or not IsValid(client)
-        or not IsValid(pnl.ply)
-        or not GetGlobalBool("ttt_locational_voice", false)
-        or pnl.ply:IsSpec()
-        or pnl.ply == client
-        or client:IsActive()
-            and pnl.ply:IsActive()
-            and (client:IsInTeam(pnl.ply) and not pnl.ply:GetSubRoleData().unknownTeam and not pnl.ply:GetSubRoleData().disabledTeamVoice and not client:GetSubRoleData().disabledTeamVoiceRecv)
-    then
-        return
-    end
-
-    local d = client:GetPos():Distance(pnl.ply:GetPos())
-
-    pnl:SetAlpha(math.max(-0.1 * d + 255, 15))
-end
-
-local PlayerVoicePanels = {}
-
 ---
 -- Called when a @{Player} starts using voice chat.
 -- @param Player ply @{Player} who started using voice chat
@@ -239,112 +313,69 @@ function GM:PlayerStartVoice(ply)
         return
     end
 
+    local plyRoleData = ply:GetSubRoleData()
+
     local client = LocalPlayer()
+    local clientTeam = client:GetTeam()
+    local clientRoleData = client:GetSubRoleData()
+
+    if ply ~= client then
+        -- if someone else than us starts talking, the UI should be updated as well
+        ply:SetSpeakingInVoice(true)
+    else
+        client[client:GetTeam() .. "_gvoice"] = not VOICE.isTeam
+
+        -- notify server this if this is a global voice chat
+        net.Start("TTT2RoleGlobalVoice")
+        net.WriteBool(not VOICE.isTeam)
+        net.SendToServer()
+    end
 
     VOICE.UpdatePlayerVoiceVolume(ply)
 
-    if not IsValid(g_VoicePanelList) or not IsValid(client) then
-        return
-    end
+    -- handle voice panel color / mode
+    local mode = VOICE_MODE_GLOBAL
 
-    -- There'd be an extra one if voice_loopback is on, so remove it.
-    GAMEMODE:PlayerEndVoice(ply, true)
-
-    -- Tell server this is global
-    if client == ply then
-        local tm = client:GetTeam()
-
-        local isGlobal = not VOICE.isTeam
-        client[tm .. "_gvoice"] = isGlobal
-
-        net.Start("TTT2RoleGlobalVoice")
-        net.WriteBool(isGlobal)
-        net.SendToServer()
-
-        VOICE.SetSpeaking(true)
-    end
-
-    local pnl = g_VoicePanelList:Add("VoiceNotify")
-    pnl:Setup(ply)
-    pnl:Dock(TOP)
-
-    local oldThink = pnl.Think
-
-    pnl.Think = function(s)
-        oldThink(s)
-
-        VoiceNotifyThink(s)
-    end
-
-    local shade = Color(0, 0, 0, 150)
-
-    -- TODO recreate all voice panels on HUD switch
-    local paintFn = function(s, w, h)
-        if not IsValid(s.ply) then
-            return
-        end
-
-        draw.RoundedBox(4, 0, 0, w, h, s.Color)
-        draw.RoundedBox(4, 1, 1, w - 2, h - 2, shade)
-    end
-
-    if huds and HUDManager then
-        local hud = huds.GetStored(HUDManager.GetHUD())
-        if hud then
-            paintFn = hud.VoicePaint or paintFn
-        end
-    end
-
-    pnl.Paint = paintFn
-
-    -- roles things
-    local tm = client:GetTeam()
-    local clrd = client:GetSubRoleData()
-
-    if
+    if ply:IsSpec() then
+        mode = VOICE_MODE_SPEC
+    elseif
         client:IsActive()
-        and tm ~= TEAM_NONE
-        and not clrd.unknownTeam
-        and not clrd.disabledTeamVoice
-        and not TEAMS[tm].alone
+        and clientTeam ~= TEAM_NONE
+        and not clientRoleData.unknownTeam
+        and not clientRoleData.disabledTeamVoice
+        and not TEAMS[clientTeam].alone
     then
         if ply == client then
-            if not client[tm .. "_gvoice"] then
-                pnl.Color = TEAMS[tm].color
+            if not client[clientTeam .. "_gvoice"] then
+                mode = VOICE_MODE_TEAM
             end
         elseif
             ply:IsInTeam(client)
-            and not (ply:GetSubRoleData().disabledTeamVoice or clrd.disabledTeamVoiceRecv)
+            and not (plyRoleData.disabledTeamVoice or clientRoleData.disabledTeamVoiceRecv)
         then
-            if not ply[tm .. "_gvoice"] then
-                pnl.Color = TEAMS[tm].color
+            if not ply[clientTeam .. "_gvoice"] then
+                mode = VOICE_MODE_TEAM
             end
         end
-    end
-
-    -- since detective (sub-) roles don't have their own team, they have a manual role color
-    -- handling here
-    if ply:IsActive() and ply:GetBaseRole() == ROLE_DETECTIVE then
-        pnl.Color = roles.DETECTIVE.color
     end
 
     ---
     -- @realm client
     -- stylua: ignore
-    pnl.Color = hook.Run("TTT2ModifyVoiceChatColor", ply, pnl.Color) or pnl.Color
+    mode = hook.Run("TTT2ModifyVoiceChatMode", ply, mode) or mode
 
-    PlayerVoicePanels[ply] = pnl
+    VOICE.SetVoiceMode(ply, mode)
 
-    local plyrd = ply:GetSubRoleData()
-
-    -- run ear gesture
+    -- add animation when player is speaking in voice
     if
         not (
             ply:IsActive()
-            and not plyrd.unknownTeam
-            and not plyrd.disabledTeamVoice
-            and not clrd.disabledTeamVoiceRecv
-        ) or (tm ~= TEAM_NONE and not TEAMS[tm].alone) and ply[tm .. "_gvoice"]
+            and not plyRoleData.unknownTeam
+            and not plyRoleData.disabledTeamVoice
+            and not clientRoleData.disabledTeamVoiceRecv
+        )
+        or (clientTeam ~= TEAM_NONE and not TEAMS[clientTeam].alone)
+            and ply[clientTeam .. "_gvoice"]
     then
         ply:AnimPerformGesture(ACT_GMOD_IN_CHAT)
     end
@@ -388,74 +419,32 @@ local function ReceiveVoiceState()
     end
 
     ply[tm .. "_gvoice"] = isGlobal
-
-    if not IsValid(PlayerVoicePanels[ply]) then
-        return
-    end
-
-    PlayerVoicePanels[ply].Color = isGlobal and VP_GREEN or (ply:GetRoleColor() or VP_RED)
 end
 net.Receive("TTT_RoleVoiceState", ReceiveVoiceState)
-
-local function VoiceClean()
-    if not PlayerVoicePanels then
-        return
-    end
-
-    for ply, pnl in pairs(PlayerVoicePanels) do
-        if IsValid(pnl) and IsValid(ply) then
-            continue
-        end
-
-        GAMEMODE:PlayerEndVoice(ply)
-    end
-end
-timer.Create("VoiceClean", 10, 0, VoiceClean)
 
 ---
 -- Called when @{Player} stops using voice chat.
 -- @param Player ply @{Player} who stopped talking
--- @param boolean no_reset whether the stored voice state shouldn't reset
 -- @hook
 -- @realm client
 -- @ref https://wiki.facepunch.com/gmod/GM:PlayerEndVoice
 -- @local
-function GM:PlayerEndVoice(ply, no_reset)
-    if IsValid(PlayerVoicePanels[ply]) then
-        PlayerVoicePanels[ply]:Remove()
-
-        PlayerVoicePanels[ply] = nil
+function GM:PlayerEndVoice(ply)
+    if not IsValid(ply) then
+        return
     end
 
-    if IsValid(ply) and not no_reset then
-        local tm = ply:GetTeam()
+    local plyTeam = ply:GetTeam()
 
-        if tm ~= TEAM_NONE and not TEAMS[tm].alone then
-            ply[tm .. "_gvoice"] = false
-        end
+    -- if someone else than us stops talking, the UI should be updated as well
+    if ply ~= LocalPlayer() then
+        ply:SetSpeakingInVoice(false)
     end
 
-    if ply == LocalPlayer() then
-        VOICE.SetSpeaking(false)
+    if plyTeam ~= TEAM_NONE and not TEAMS[plyTeam].alone then
+        ply[plyTeam .. "_gvoice"] = false
     end
 end
-
-local function CreateVoiceVGUI()
-    g_VoicePanelList = vgui.Create("DPanel")
-    g_VoicePanelList:ParentToHUD()
-    g_VoicePanelList:SetPos(25, 25)
-    g_VoicePanelList:SetSize(200, ScrH() - 200)
-    g_VoicePanelList:SetPaintBackground(false)
-
-    MutedState = vgui.Create("DLabel")
-    MutedState:SetPos(ScrW() - 200, ScrH() - 50)
-    MutedState:SetSize(200, 50)
-    MutedState:SetFont("Trebuchet18")
-    MutedState:SetText("")
-    MutedState:SetTextColor(Color(240, 240, 240, 250))
-    MutedState:SetVisible(false)
-end
-hook.Add("InitPostEntity", "CreateVoiceVGUI", CreateVoiceVGUI)
 
 --local MuteStates = {MUTE_NONE, MUTE_TERROR, MUTE_ALL, MUTE_SPEC}
 
@@ -732,50 +721,12 @@ function VOICE.CanSpeak()
     return client.voice_battery > VOICE.battery_min or IsRoleChatting(client)
 end
 
-local speaker = surface.GetTextureID("voice/icntlk_sv")
-
----
--- Draws a popup displaying the speaking @{Player}s
--- @param Player client This should be the local @{Player}
--- @return boolean
--- @realm client
--- @internal
-function VOICE.Draw(client)
-    local b = client.voice_battery
-
-    if not b or not VOICE.battery_max or b >= VOICE.battery_max then
-        return
-    end
-
-    local x, y = 25, 10
-    local w, h = 200, 6
-
-    if b < VOICE.battery_min and CurTime() % 0.2 < 0.1 then
-        surface.SetDrawColor(200, 0, 0, 155)
-    else
-        surface.SetDrawColor(0, 200, 0, 255)
-    end
-
-    surface.DrawOutlinedRect(x, y, w, h)
-
-    surface.SetTexture(speaker)
-    surface.DrawTexturedRect(5, 5, 16, 16)
-
-    x = x + 1
-    y = y + 1
-    w = w - 2
-    h = h - 2
-
-    surface.SetDrawColor(0, 200, 0, 150)
-    surface.DrawRect(x, y, w * math.Clamp((client.voice_battery - 10) / 90, 0, 1), h)
-end
-
 ---
 -- This hook can be used to modify the background color of the voice chat
--- box that is rendered on the client.
+-- box that is rendered on the client. This is done by setting the voice chat mode.
 -- @param ply The player that started a voice chat
--- @param Color clr The color that is used if this hook does not modify it
--- @return Color The new and modified color
+-- @param number mode The voice chat mode that is used if this hook does not modify it
+-- @return number The new and modified mode
 -- @hook
 -- @realm client
-function GM:TTT2ModifyVoiceChatColor(ply, clr) end
+function GM:TTT2ModifyVoiceChatMode(ply, mode) end